🐛Fix: generateMetadata 오류 수정
generateMetadata 서버 컴포넌트에서 사용 가능하므로 layout.tsx에서 하도록 수정
