+++ src/app/(authorized)/(trip)/trip/[id]/page.tsx
@@ -6,37 +6,6 @@ import NavTitle from '@ui/common/NavTitle';
 import TripTask from '@ui/trip/TripTask';
 import TripJoinAccess from '@ui/trip/TripJoinAccess';
 import { useGetTrip } from '@hooks/trip/useGetTrip';
-import { Metadata } from 'next';
-import { getTrip } from '@lib/api/service/trip.api';
-
-export async function generateMetadata({
-  params,
-}: {
-  params: { id: string };
-}): Promise<Metadata> {
-  const id = Number(params.id);
-  const tripInfo = await getTrip(id);
-  const pageTitle = tripInfo?.result?.name || '여행 상세';
-
-  return {
-    title: pageTitle,
-    openGraph: {
-      title: pageTitle,
-      url: `https://torip.com/trip/${params.id}`,
-      siteName: 'Torip',
-      images: [
-        {
-          url: '/asset/image/logo.png',
-          width: 1200,
-          height: 630,
-          alt: 'Torip Open Graph Image',
-        },
-      ],
-      locale: 'ko_KR',
-      type: 'website',
-    },
-  };
-}
 
 export default function Trip({ params }: { params: { id: string } }) {
   const { id } = params;
