--- src/app/(authorized)/(trip)/trip/layout.tsx
@@ -1,10 +1,0 @@
-interface ITripLayoutProps {
-  params: { id: string };
-  children: React.ReactNode;
-}
-
-export default async function TripLayout({ children }: ITripLayoutProps) {
-  return (
-    <div className="flex min-h-full flex-1 flex-col gap-6">{children}</div>
-  );
-}
